requirements.txt
requests==2.28.1
beautifulsoup4==4.11.1
scikit-learn==1.2.0
numpy==1.24.1
pandas==1.5.3
shodan==1.28.0
dnspython==2.2.1
nmap==0.7.1
boto3==1.26.0
google-cloud-compute==1.10.0
azure-identity==1.12.0
azure-mgmt-compute==29.0.0
torch==1.13.1
transformers==4.25.1
cryptography==39.0.0
pytest==7.2.0
reportlab==3.6.12
sphinx==5.3.0
flake8==6.0.0
